+++ contracts/test/PoolMock.sol
@@ -16,15 +16,15 @@ contract PoolMock is Pool {
 
     function formatTokenId(
         address operator,
-        uint64 lower,
-        uint64 upper
+        uint256 lower,
+        uint256 upper
     ) external pure returns (uint256 tokenId) {
         return PoolStorage.formatTokenId(operator, lower, upper);
     }
 
     function parseTokenId(
         uint256 tokenId
-    ) external pure returns (address operator, uint64 lower, uint64 upper) {
+    ) external pure returns (address operator, uint256 lower, uint256 upper) {
         return PoolStorage.parseTokenId(tokenId);
     }
 }
